+++ CHANGELOG.md
@@ -73,7 +73,6 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Button SCSS variables related to ghost and link types, and paddings based on size
 
-
 ## 07/10/2022
 
 ### Changed
@@ -86,10 +85,18 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Separate RadioGroup component
 
-
 ## 08/10/2022
 
 ### Changed
 
 - Refactor Checkbox & CheckboxGroup components
 
+## 15/10/2022
+
+### Added
+
+- onDelete & size props to the Chip component
+
+### Changed
+
+- Markup and SCSS of Chip component